My Player Class
The character is a 2D wolf sprite I made in Photoshop and he is seen from a side view. He jumps with spacebar and the up arrow key. He is animated and has 3 poses - IDLE, RUNNING, JUMPING. Vector geometry is used to add x and y components together to move the wolf. The player character is the only thing that moves.